requirements.txt
mattermostdriver>=7.3.0
python-docx>=0.8.11
PyPDF2>=3.0.1
openpyxl>=3.1.2
striprtf>=0.0.26
requests>=2.31.0
openai>=1.12.0
reportlab>=4.0.7
python-dotenv>=1.0.0
beautifulsoup4>=4.12.2
lxml>=4.9.3
json5>=0.9.14 